+++ common/src/main/kotlin/com/cobblemon/mod/common/api/storage/player/PlayerData.kt
@@ -10,8 +10,8 @@ package com.cobblemon.mod.common.api.storage.player
 
 import com.cobblemon.mod.common.Cobblemon
 import com.cobblemon.mod.common.CobblemonNetwork.sendPacket
+import com.cobblemon.mod.common.CobblemonSounds
 import com.cobblemon.mod.common.net.messages.client.starter.SetClientPlayerDataPacket
-import com.cobblemon.mod.common.util.cobblemonResource
 import java.util.UUID
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.util.Identifier
@@ -23,6 +23,7 @@ data class PlayerData(
     var starterSelected: Boolean,
     var starterUUID: UUID?,
     var keyItems: MutableSet<Identifier>,
+    var battleTheme: Identifier?,
     val extraData: MutableMap<String, PlayerDataExtension>,
 ) {
     var advancementData: PlayerAdvancementData = PlayerAdvancementData()
@@ -40,6 +41,7 @@ data class PlayerData(
             starterSelected =  false,
             starterUUID =  null,
             keyItems = mutableSetOf(),
+            battleTheme = null,
             extraData = mutableMapOf()
         )
     }